- Fixed the NR recording subsystem so that the sampling frequency is recorded as a double instead of an int32
git-svn-id: https://neurorighter.googlecode.com/svn/trunk@537 471701b0-e017-11dd-9759-139c9ab14fc6

+++ NeuroRighter/FileWriting/FileOutputRemapped.cs
@@ -9,7 +9,7 @@ namespace NeuroRighter.FileWriting
 {
     internal class FileOutputRemapped : FileOutput
     {
-        internal FileOutputRemapped(string filenameBase, int numChannels, int samplingRate, int fileType, Task recordingTask, string extension, double preampgain) :
+        internal FileOutputRemapped(string filenameBase, int numChannels, double samplingRate, int fileType, Task recordingTask, string extension, double preampgain) :
             base( filenameBase,  numChannels, samplingRate, fileType,  recordingTask, extension, preampgain) {}
 
         internal override void read(short[,] data, int numChannelsData, int startChannelData, int length)
